+++ src/include/smash/configuration.h
@@ -686,11 +686,14 @@ class Configuration {
           s.set(IncludedMultiParticleReactions::Meson_3to1);
         } else if (x == "Deuteron_3to2") {
           s.set(IncludedMultiParticleReactions::Deuteron_3to2);
+        } else if (x == "NNbar_5to2") {
+          s.set(IncludedMultiParticleReactions::NNbar_5to2);
         } else {
           throw IncorrectTypeInAssignment(
               "The value for key \"" + std::string(key_) +
-              "\" should be \"All\", \"Meson_3to1\" or "
-              "\"Deuteron_3to2\", or any combination of these.");
+              "\" should be \"All\", \"Meson_3to1\", "
+              "\"Deuteron_3to2\" or \"NNbar_5to2\", or any combination of "
+              "these.");
         }
       }
       return s;
@@ -927,12 +930,16 @@ class Configuration {
       if (s == "resonances") {
         return NNbarTreatment::Resonances;
       }
+      if (s == "two to five") {
+        return NNbarTreatment::TwoToFive;
+      }
       if (s == "strings") {
         return NNbarTreatment::Strings;
       }
       throw IncorrectTypeInAssignment(
           "The value for key \"" + std::string(key_) + "\" should be " +
-          "\"no annihilation\", \"detailed balance\", or \"strings\".");
+          "\"no annihilation\", \"resonances\", \"two to five\" or " +
+          " \"strings\".");
     }
 
     /**